Feature [BPT1-58] syaiful: create, update, delete item test (#28)
* try to solve multer

* feat [BPT1-58] syaiful: fix create, update, delete item

* fix: solve conflicts

+++ libs/media.handling.js
@@ -4,20 +4,29 @@ const multer = require("multer");
 const cloudinary = require("cloudinary").v2;
 const fs = require("fs");
 
-cloudinary.config({
-  secure: true,
-});
+let uploadCloud = async (file) => {}
 
-const uploadCloud = async (file) => {
-  try {
-    const result = await cloudinary.uploader.upload(file);
-    return result.secure_url;
-  } catch (error) {
-    console.log(error);
-  } finally {
-    fs.unlinkSync(file);
-  }
-};
+if (process.env.NODE_ENV === "test") {
+  cloudinary.config({
+    secure: true,
+  });
+  
+  uploadCloud = async (file) => {
+    try {
+      const result = await cloudinary.uploader.upload(file);
+      return result.secure_url;
+    } catch (error) {
+      console.log(error);
+    } finally {
+      fs.unlinkSync(file);
+    }
+  };
+
+} else {
+  uploadCloud = async file => ({
+    secure_url: "https://cloudinary/" + file
+  })
+}
 
 const localStore = multer.diskStorage({
   destination: function (req, file, cb) {
